+++ app/database/migrations/2022_09_02_012212_add_highest_improvement_total_to_dominions_table.php
@@ -1,0 +1,32 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class AddHighestImprovementTotalToDominionsTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('dominions', function (Blueprint $table) {
+            $table->unsignedInteger('highest_improvement_total')->after('highest_land_achieved')->default(0);
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('dominions', function (Blueprint $table) {
+            $table->dropColumn('highest_improvement_total');
+        });
+    }
+}