+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/). 
 - Fireball damage increased to 2.75% (from 2.65%)
 - Sabotage damage increased to 2.5% (from 2.4%)
 - Cyclone cost reduced to 2x (from 3x)
+- Castle improvements destroyed by Lightning Bolt can now be 'repaired' (investment is doubled)
 - Trick of the Light tech now protects mana instead of decreasing mana cost
 - Reinstated prestige penalty for invasions against bots (-5% per hit after the 4th, max -50%)
 - Dark Elf Spellblade: +0.5 DP (from 2)